Cleanup logic for computed attributes
